Implemented Screenshot and save to file function
instead of a pause and play button, we take a screenshot of the current animation from renderer send it to main using ipc and download it. next step is to replace the screen with the screenshot until the user presses on another button to resume live input. this should be done through manipulating the html and css and without stopping or interfering the javascript code of the animation